+++ packages/providers/src/index.ts
@@ -14,6 +14,7 @@ export * from "./aniliberty-streaming/index.js";
 export * from "./filmix-streaming/index.js";
 export * from "./veoveo-streaming/index.js";
 export * from "./videohub-streaming/index.js";
+export * from "./rutube-streaming/index.js";
 export * from "./yts-torrent/index.js";
 export * from "./jacred-torrent/index.js";
 export * from "./bitsearch-torrent/index.js";